CEDS-1204 Move RadioOption into fields package
