+++ app/views/choice_page.scala.html
@@ -19,7 +19,7 @@
 @import forms.Choice
 @import forms.Choice.AllowedChoiceValues
 @import uk.gov.hmrc.play.views.html._
-@import views.components.RadioOption
+@import views.components.fields.RadioOption
 
 @this(main_template: views.html.main_template)
