+++ app/views/components/input_radio.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import views.components.RadioOption
+@import views.components.fields.RadioOption
 @(
     field: Field,
     legend: String,
