+++ app/views/consignment_references.scala.html
@@ -19,7 +19,7 @@
 @import uk.gov.hmrc.play.views.html._
 @import forms.ConsignmentReferences
 @import forms.ConsignmentReferences.AllowedReferences._
-@import views.components.RadioOption
+@import views.components.fields.RadioOption
 
 @this(main_template: views.html.main_template)
