+++ app/views/goods_departed.scala.html
@@ -19,7 +19,7 @@
 @import forms.GoodsDeparted.AllowedPlaces._
 @import uk.gov.hmrc.play.views.html._
 @import forms.GoodsDeparted
-@import views.components.RadioOption
+@import views.components.fields.RadioOption
 
 @this(main_template: views.html.main_template)
