+++ app/views/transport.scala.html
@@ -19,7 +19,7 @@
 @import uk.gov.hmrc.play.views.html._
 @import forms.Transport
 @import forms.Transport.ModesOfTransport._
-@import views.components.RadioOption
+@import views.components.fields.RadioOption
 
 @this(main_template: views.html.main_template)
